+++ libs/testing-utils/src/vivarium/testing_utils/automated_validation/__init__.py
@@ -1,1 +1,0 @@
-from vivarium.testing_utils.automated_validation.interface import ValidationContext